+++ packages/ws-worker/src/server.ts
@@ -202,6 +202,15 @@ function createServer(engine: RuntimeEngine, options: ServerOptions = {}) {
           input,
         } = await joinRunChannel(app.socket, token, id, logger);
 
+        // Setup collections
+        if (plan.workflow.credentials?.collections_token) {
+          plan.workflow.credentials.collections_token = token;
+        }
+        if (plan.workflow.credentials?.collections_endpoint) {
+          plan.workflow.credentials.collections_endpoint =
+            app.options.lightning;
+        }
+
         // Default the payload limit if it's not otherwise set on the run options
         if (!('payloadLimitMb' in options)) {
           options.payloadLimitMb = app.options.payloadLimitMb;